fix: move the hooks from ramp folder
